+++ JotTouchExample/JotTouchExample/AdonitSDK.framework/Versions/A/Headers/JotStroke.h
@@ -40,12 +40,25 @@
 */
 @property (nonatomic, readonly) NSUInteger tapCount;
 
-/*! The pressure associated with the jotStroke.
+/*! A float between 0 and 1 indicating the pressure associated with the jotStroke, calculated from the raw pressure.
  */
-@property (nonatomic, readonly) NSUInteger pressure;
+@property (nonatomic, readonly) CGFloat pressure;
 
-/*! An array of coalescedJotStrokes similar to the concept of coalesced UITouches. If you do not enable coalescedJotStrokes, this will return an array populated by this jotStroke instead. To enable coalescedJotStrokes turn on "coalescedJotStrokesEnabled" on an instance of JotStylusManager. Note: Coalesced strokes are snapshots, and each instance is a separate object. For tracking a persistent stroke across events, use the non-coalesced parent jotStroke instead of strokes from this array. */
+/*! The raw pressure associated with the jotStroke.
+ */
+@property (nonatomic, readonly) NSUInteger rawPressure;
+
+/*! A value of 0 radians indicates that the stylus is parallel to the surface; when the stylus is perpendicular to the surface, altitudeAngle is Pi/2.
+    Styluses that do not support altitude angle (Such as Jot Script and Jot Touch 4) will return a value of Pi/4.
+ */
+@property (nonatomic, readonly) CGFloat altitudeAngle;
+
+/*! An array of coalescedJotStrokes similar to the concept of coalesced UITouches. If you do not enable coalescedJotStrokes, this will return an array populated by this jotStroke instead. To enable coalescedJotStrokes turn on "coalescedJotStrokesEnabled" on an instance of JotStylusManager. Note: Coalesced strokes are snapshots, and each instance is a separate object. For tracking a persistent stroke across events, use the non-coalesced parent jotStroke instead of strokes from this array. 
+ */
 @property (nonatomic, readonly) NSArray *coalescedJotStrokes;
 
+/*! An array of auxiliary JotStrokes for stroke events that are predicted to occur for a given main JotStroke. These predictions may not exactly match the real behavior of the stroke as the stylus moves, so they should be interpreted as an estimate. To enable predictedJotStrokes turn on "predictedJotStrokesEnabled" on an instance of JotStylusManager.
+  */
+@property (nonatomic, readonly) NSArray *predictedJotStrokes;
 
 @end